Extract Item Stats Card Into a Separate Component
